+++ web/src/modules/Index/index.tsx
@@ -12,7 +12,6 @@ import WidgetWeather from "./components/WidgetWeather";
 import WidgetTodo from "./components/WidgetToDo";
 import WidgetClock from "./components/WidgetClock";
 import { IS_SELF_HOST } from "@utils/resources/isSelfHost";
-import { useEffect } from "react";
 
 const PageIndex = () => {
   const t = useTranslate();
@@ -47,12 +46,6 @@ const PageIndex = () => {
 
   // Change settings from params
   useSettingsParams();
-
-  useEffect(() => {
-    fetch("/api/ip");
-
-    fetch(`${process.env.API_URL_EU1}/ip`);
-  }, []);
 
   return (
     <Container