+++ web/src/modules/Search/components/components/SearchSection/components/SearchSectionInput/index.tsx
@@ -1,5 +1,5 @@
 import { ActionIcon, Combobox, Divider, Flex, Image, TextInput, useCombobox } from "@mantine/core";
-import { IconSearch,  IconX } from "@tabler/icons-react";
+import { IconSearch, IconSparkles, IconX } from "@tabler/icons-react";
 import { getIconStyle } from "@utils/functions/iconStyle";
 import { useEffect, useRef, useState } from "react";
 
@@ -26,6 +26,7 @@ const SearchSectionInput = () => {
 
   const useAutocomplete = useSettingsStore((state) => state.useAutocomplete);
   const privateSearch = useSettingsStore((state) => state.privateSearch);
+  const useAIAnswers = useSettingsStore((state) => state.useAIAnswers);
 
   const searchQuery = useSearchStore((state) => state.searchQuery);
   const setSearchQuery = useSearchStore((state) => state.setSearchQuery);
@@ -36,7 +37,7 @@ const SearchSectionInput = () => {
   // Autocomplete API
   const { data, trigger, reset } = useAutocompleteSWR();
 
-  const handleSearch = (query: string) => {
+  const handleSearch = (query: string, withAI: boolean) => {
     const paramsTab = searchParams.get("tab") || "general";
     const paramsQ = searchParams.get("q") || "";
 
@@ -49,12 +50,17 @@ const SearchSectionInput = () => {
     // Unfocus input on search
     inputRef.current?.blur();
 
+    const qParam = `q=${encodeURIComponent(query)}`;
+    const tabParam = `tab=${tabFromSyntax || paramsTab}`;
+    const aiParam = withAI ? "&ai=1" : "";
+
+    // Handle Private Search
     // Handle Private Search
     if (privateSearch) {
       setSearchQuery(encodeURIComponent(query));
-      return navigate(`/search?tab=${tabFromSyntax || paramsTab}`);
+      return navigate(`/search?${tabParam}${aiParam}`);
     }
-    navigate(`/search?q=${encodeURIComponent(query)}&tab=${tabFromSyntax || paramsTab}`);
+    navigate(`/search?${qParam}&${tabParam}${aiParam}`);
   };
 
   const handleClear = () => {
@@ -84,7 +90,7 @@ const SearchSectionInput = () => {
   return (
     <Combobox
       onOptionSubmit={(val) => {
-        handleSearch(val);
+        handleSearch(val, false);
         combobox.closeDropdown();
       }}
       resetSelectionOnOptionHover
@@ -110,42 +116,34 @@ const SearchSectionInput = () => {
             const isSubmitOption = combobox.getSelectedOptionIndex() !== -1;
 
             if (e.key === "Enter" && !isSubmitOption) {
-              handleSearch(q);
+              handleSearch(q, false);
             }
           }}
-          // leftSection={
-          //   isSm && (
-          //     <RemixLink className={classes.app_logo_mobile} to="/">
-          //       {/* <IconTriangleFilled style={getIconStyle(22)} /> */}
-          //       <Image src="/assets/logo.svg" w={22} h={22} />
-          //     </RemixLink>
-          //   )
-          // }
           rightSection={
-            <Flex align="center" justify="flex-end" gap={4}>
+            <Flex align="center" justify="flex-end">
               {q.length >= 1 && (
                 <>
                   <ActionIcon
                     size="lg"
                     radius="sm"
-                    color="gray.8"
-                    variant="transparent"
+                    color="gray"
+                    variant="subtle"
                     onClick={handleClear}
                   >
                     <IconX style={getIconStyle(20)} stroke={1.5} />
                   </ActionIcon>
 
-                  <Divider orientation="vertical" w={1} my={9} color="gray.7" />
+                  <Divider orientation="vertical" w={1} my={9} mx={4} color="gray.7" />
                 </>
               )}
 
               <ActionIcon
-                w={40}
-                h={40}
-                radius="md"
-                color="gray.4"
-                variant="transparent"
-                onClick={() => handleSearch(q)}
+                size="lg"
+                radius="sm"
+                color="blue"
+                variant="subtle"
+                onClick={() => handleSearch(q, false)}
+                mr={4}
               >
                 <IconSearch
                   style={getIconStyle(20)}
@@ -153,9 +151,27 @@ const SearchSectionInput = () => {
                   // color="white"
                 />
               </ActionIcon>
+
+              {useAIAnswers && (
+                <ActionIcon
+                  size="lg"
+                  radius="sm"
+                  color="pink"
+                  variant="subtle"
+                  onClick={() => handleSearch(q, true)}
+                  mr={4}
+                >
+                  <IconSparkles
+                    style={getIconStyle(20)}
+                    stroke={2}
+                    // color="white"
+                  />
+                </ActionIcon>
+              )}
             </Flex>
           }
-          rightSectionWidth={q.length >= 1 ? 83 : 40}
+          // rightSectionWidth={q.length >= 1 ? 83 : 40}
+          rightSectionWidth="fit-content"
           // onClick={() => combobox.openDropdown()}
           onFocus={() => combobox.openDropdown()}
           onBlur={() => combobox.closeDropdown()}